:construction: mybatis 返回影响条数为Integer类型

+++ src/test/java/com/dddinjava/portal/dao/AccountMapperTest.java
@@ -20,7 +20,7 @@ class AccountMapperTest {
     @DisplayName("保存")
     void save() {
         Account account = new Account("dddinjava", "123", "dddinjava");
-        Long count = accountMapper.save(account);
+        Integer count = accountMapper.save(account);
         assertEquals(1, count);
     }
